refactor: textfield showcases and storybook
